+++ linking/cid/cidLink.go
@@ -1,18 +1,16 @@
 package cidlink
 
 import (
-	"bytes"
-	"context"
 	"fmt"
-	"io"
 
 	cid "github.com/ipfs/go-cid"
 	ipld "github.com/ipld/go-ipld-prime"
+	multihash "github.com/multiformats/go-multihash"
 )
 
 var (
-	_ ipld.Link        = Link{}
-	_ ipld.LinkBuilder = LinkBuilder{}
+	_ ipld.Link          = Link{}
+	_ ipld.LinkPrototype = LinkPrototype{}
 )
 
 // Link implements the ipld.Link interface using a CID.
@@ -26,92 +24,50 @@ type Link struct {
 	cid.Cid
 }
 
-// byteAccessor is a reader interface that can access underlying bytes
-type byteAccesor interface {
-	Bytes() []byte
-}
-
-func (lnk Link) Load(ctx context.Context, lnkCtx ipld.LinkContext, na ipld.NodeAssembler, loader ipld.Loader) error {
-	// Open the byte reader.
-	r, err := loader(lnk, lnkCtx)
-	if err != nil {
-		return err
-	}
-	// Tee into hash checking and unmarshalling.
-	mcDecoder, exists := multicodecDecodeTable[lnk.Prefix().Codec]
-	if !exists {
-		return fmt.Errorf("no decoder registered for multicodec %d", lnk.Prefix().Codec)
-	}
-	var hasherBytes []byte
-	var decodeErr error
-	byteBuf, ok := r.(byteAccesor)
-	if ok {
-		hasherBytes = byteBuf.Bytes()
-		decodeErr = mcDecoder(na, r)
-	} else {
-		var hasher bytes.Buffer // multihash only exports bulk use, which is... really inefficient and should be fixed.
-		decodeErr = mcDecoder(na, io.TeeReader(r, &hasher))
-		// Error checking order here is tricky.
-		//  If decoding errored out, we should still run the reader to the end, to check the hash.
-		//  (We still don't implement this by running the hash to the end first, because that would increase the high-water memory requirement.)
-		//   ((Which we experience right now anyway because multihash's interface is silly, but we're acting as if that's fixed or will be soon.))
-		//  If the hash is rejected, we should return that error (and even if there was a decodeErr, it becomes irrelevant).
-		if decodeErr != nil {
-			_, err := io.Copy(&hasher, r)
-			if err != nil {
-				return err
-			}
-		}
-		hasherBytes = hasher.Bytes()
-	}
-
-	cid, err := lnk.Prefix().Sum(hasherBytes)
-	if err != nil {
-		return err
-	}
-	if cid != lnk.Cid {
-		return fmt.Errorf("hash mismatch!  %q (actual) != %q (expected)", cid, lnk.Cid)
-	}
-	if decodeErr != nil {
-		return decodeErr
-	}
-	return nil
-}
-func (lnk Link) LinkBuilder() ipld.LinkBuilder {
-	return LinkBuilder{lnk.Cid.Prefix()}
+func (lnk Link) Prototype() ipld.LinkPrototype {
+	return LinkPrototype{lnk.Cid.Prefix()}
 }
 func (lnk Link) String() string {
 	return lnk.Cid.String()
 }
 
-type LinkBuilder struct {
+type LinkPrototype struct {
 	cid.Prefix
 }
 
-func (lb LinkBuilder) Build(ctx context.Context, lnkCtx ipld.LinkContext, node ipld.Node, storer ipld.Storer) (ipld.Link, error) {
-	// Open the byte writer.
-	w, commit, err := storer(lnkCtx)
+func (lp LinkPrototype) BuildLink(hashsum []byte) ipld.Link {
+	// Does this method body look surprisingly complex?  I agree.
+	//  We actually have to do all this work.  The go-cid package doesn't expose a constructor that just lets us directly set the bytes and the prefix numbers next to each other.
+	//  No, `cid.Prefix.Sum` is not the method you are looking for: that expects the whole data body.
+	//  Most of the logic here is the same as the body of `cid.Prefix.Sum`; we just couldn't get at the relevant parts without copypasta.
+	//  There is also some logic that's sort of folded in from the go-multihash module.  This is really a mess.
+	//  The go-cid package needs review.  So does go-multihash.  Their responsibilies are not well compartmentalized and they don't play well with other stdlib golang interfaces.
+	p := lp.Prefix
+
+	length := p.MhLength
+	if p.MhType == multihash.ID {
+		length = -1
+	}
+	if p.Version == 0 && (p.MhType != multihash.SHA2_256 ||
+		(p.MhLength != 32 && p.MhLength != -1)) {
+		panic(fmt.Errorf("invalid cid v0 prefix"))
+	}
+
+	if length != -1 {
+		hashsum = hashsum[:p.MhLength]
+	}
+
+	mh, err := multihash.Encode(hashsum, p.MhType)
 	if err != nil {
-		return nil, err
+		panic(err) // No longer possible, but multihash still returns an error for legacy reasons.
 	}
-	// Marshal, teeing into the storage writer and the hasher.
-	mcEncoder, exists := multicodecEncodeTable[lb.Prefix.Codec]
-	if !exists {
-		return nil, fmt.Errorf("no encoder registered for multicodec %d", lb.Prefix.Codec)
+
+	switch lp.Prefix.Version {
+	case 0:
+		return Link{cid.NewCidV0(mh)}
+	case 1:
+		return Link{cid.NewCidV1(p.Codec, mh)}
+	default:
+		panic(fmt.Errorf("invalid cid version"))
 	}
-	var hasher bytes.Buffer // multihash-via-cid only exports bulk use, which is... really inefficient and should be fixed.
-	w = io.MultiWriter(&hasher, w)
-	err = mcEncoder(node, w)
-	if err != nil {
-		return nil, err
-	}
-	cid, err := lb.Prefix.Sum(hasher.Bytes())
-	if err != nil {
-		return nil, err
-	}
-	lnk := Link{cid}
-	if err := commit(lnk); err != nil {
-		return lnk, err
-	}
-	return lnk, nil
 }